Initial commit
The basic strategy is to

 1. Choose a default evil state for magit buffers
 2. Pick the right magit buffers for this state
 3. Make the magit keymaps overriding maps for those buffers
 4. Use `evil-define-key` to set new bindings in auxiliary maps
 5. Make minor changes to the section maps to make everything work.

There are essentially two active keymaps for magit at the same time, the
one that magit starts with (always available through emacs-state) and
the ones provided by evil-magit (available in the correct evil-state).
Therefore, there is no need for this to be a minor mode that gets
enabled and disabled.