Introduce and use evil-magit-define-key
The fancy stuff happening in `evil-define-key` with delayed execution
was causing problems for some users and is completely unnecessary in
this case, since all of the keymaps should be initialized already.
`evil-define-key` being a macro is also annoying. So introduce the
simpler `evil-magit-define-key` to do the work of putting keys in the
evil auxiliary maps.