Make more magit commands available in visual state
Previously some useful magit commands were shadowed in visual state like
A for cherry picking a selection of commits from a log buffer. This
commit attempts to fix all of these by only making movement commands
similar to the one available in normal state available in visual state
and letting the rest of the magit commands be available after that.

This only affects those who use evil-magit-use-y-for-yank.